# docs/APIv1/API.yml
openapi: '3.0.2'
info:
  title: Transition-API
  version: '1.1'
servers:
  - url: http://localhost:8080/

paths:
  /token:
    post:
      summary: Retrieve authentication token for specified user
      description: Returns an authentication token. This token is required in the Authentication header of requests to other endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'token.yml#/tokenRequest'
      responses:
        '200':
          description: Returns the generated token for specified user
          content:
            text/plain:
              schema:
                $ref: 'token.yml#/tokenResponseSuccess'
        '400':
          description: User credentials are missing
          content:
            text/plain:
              schema:
                $ref: 'token.yml#/tokenResponseBadRequest'
        '401':
          description: User credentials are invalid
          content:
            text/plain:
              schema:
                $ref: 'token.yml#/tokenResponseUnauthorized'

  /api/versions:
    get:
      summary: Get the versions supported by this API
      description: Get the latest version of the API supported by this version of Transition
      x-added-in-version: 1.1
      responses:
        '200':
          description: Returns a list of the latest version for each supported major version of the API
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  description: Version number
                example: ["1.1"]

  /api/v1/paths:
    get:
      summary: Get all paths
      description: Get all paths which are currently loaded in Transition
      responses:
        '200':
          description: Returns all paths currently loaded in Transition as a GeoJSON LineString FeatureCollection object
          content:
            application/json:
              schema:
                $ref: 'paths.yml#/paths'

  /api/v1/nodes:
    get:
      summary: Get all nodes
      description: Get all nodes which are currently loaded in Transition
      responses:
        '200':
          description: Returns all nodes currently loaded in Transition as a GeoJSON Point FeatureCollection object
          content:
            application/json:
              schema:
                $ref: 'nodes.yml#/nodes'
  
  /api/v1/scenarios:
    get:
      summary: Get all scenarios
      description: Get all scenarios which are currently loaded in Transition
      responses:
        '200':
          description: Returns a list of all scenarios currently loaded in Transition with their parameters
          content:
            application/json:
              schema:
                $ref: 'scenarios.yml#/scenarios'

  /api/v1/agencies:
    get:
      summary: Get all agencies
      description: Get all agencies which are currently available in Transition
      x-added-in-version: 1.1
      responses:
        '200':
          description: Returns a list of all agencies currently available in Transition
          content:
            application/json:
              schema:
                $ref: 'agencies.yml#/agencies'

  /api/v1/services:
    get:
      summary: Get all services
      description: Get all services which are currently available in Transition
      x-added-in-version: 1.1
      responses:
        '200':
          description: Returns a list of all services currently available in Transition
          content:
            application/json:
              schema:
                $ref: 'services.yml#/services'

  /api/v1/lines:
    get:
      summary: Get all lines
      description: Get all lines which are currently available in Transition
      x-added-in-version: 1.1
      responses:
        '200':
          description: Returns a list of all lines currently available in Transition
          content:
            application/json:
              schema:
                $ref: 'lines.yml#/lines'

  /api/v1/routing-modes:
    get: 
      summary: Get all routing modes
      description: Get all routing modes which are currently availible in Transition
      responses:
        '200':
          description: Returns all currently availible routing modes in Transition as a list of strings
          content:
            application/json:
              schema:
                $ref: 'routingModes.yml#/routingModes'
  
  /api/v1/summary:
    post:
      summary: Get summary of route calculation result
      description: Return a summary of each transit object of a certain type used by the route calculation result
      x-added-in-version: 1.1
      parameters:
      - $ref: "summary.yml#/originParam"
      - $ref: "summary.yml#/destinationParam"
      - $ref: "summary.yml#/scenarioParam"
      - $ref: "summary.yml#/departureTimeParam"
      - $ref: "summary.yml#/arrivalTimeParam"
      - $ref: "summary.yml#/alternativesParam"
      - $ref: "summary.yml#/minWaitingTimeParam"
      - $ref: "summary.yml#/maxAccessTravelTimeParam"
      - $ref: "summary.yml#/maxEgressTravelTimeParam"
      - $ref: "summary.yml#/maxTransferTravelTimeParam"
      - $ref: "summary.yml#/maxTravelTimeParam"
      - $ref: "summary.yml#/maxFirstWaitingTime"
      - in: query
        name: type
        schema:
          type: string
          enum: [line]
        required: true
        description: The type of objects for which to get the summary
      responses:
        '200':
          description: Successful query, but may not have returned a routing result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: 'summary.yml#/data_error'
                  - $ref: 'summary.yml#/successResponse'
                discriminator:
                  propertyName: status
                  mapping:
                    success: 'summary.yml#/successResponse'
                    data_error: 'summary.yml#/data_error'
        '400':
          description: Query parameters are invalid
          content:
            application/json:
              schema:
                $ref: 'summary.yml#/query_error'

  /api/v1/route:
    post:
      summary: Calculate route between two locations
      description: Calculate a route between a given origin and destination with the provided modes of transport
      parameters:
        - in: query
          name: withGeojson
          schema:
            type: boolean
          description: Whether or not to also provide the calculated routes in GeoJSON format in the response. The default value is true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'route.yml#/routeRequest'
      responses:
        '200':
          description: Successful query, returns a routing result
          content:
            application/json:
              schema:
                $ref: 'route.yml#/routeResponseSuccess'
        '400':
          description: Query parameters are invalid
          content:
            application/json:
              schema:
                $ref: 'route.yml#/routeResponseBadRequest'

  /api/v1/accessibility:
    post: 
      summary: Calculate accessibility map
      description: Calculate a map of all accessible locations within a specified time frame from or to a given location
      parameters:
        - in: query
          name: withGeojson
          schema:
            type: boolean
          description: Whether or not to also provide the accessibility map in GeoJSON format in the response. Otherwise, only a list of accessible nodes will be provided. The default value is true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'accessibilityMap.yml#/accessibilityMapRequest'
      responses:
        '200':
          description: Successful query, returns an accessibility map
          content:
            application/json:
              schema:
                $ref: 'accessibilityMap.yml#/accessibilityMapResponseSuccess'
        '400':
          description: Query parameters are invalid
          content:
            application/json:
              schema:
                $ref: 'accessibilityMap.yml#/accessibilityMapResponseBadRequest'